+++ lib/commands/8ball.ts
@@ -37,7 +37,7 @@ export const command = new Classes.Command({
 	}, //body
 });
 
-export async function init(vale: Classes.Vale) {
+export async function init(vale: Classes.Vale): Promise<Classes.Command> {
 	command.usage = vale.opts.config.prefix + command.usage;
 	command.exp = new RegExp('^' + vale.opts.config.prefix + "8(ball)?( .+)?$", "smi");
 	Classes.fetch("https://nekos.life/api/v2/8ball").then((reply: string) => command.data.cache.push(reply));