+++ lib/commands/yomomma.ts
@@ -37,7 +37,7 @@ export const command = new Classes.Command({
 	}, //body
 });
 
-export async function init(vale: Classes.Vale) {
+export async function init(vale: Classes.Vale): Promise<Classes.Command> {
 	command.usage = vale.opts.config.prefix + command.usage;
 	command.exp = new RegExp('^' + vale.opts.config.prefix + "yo(mm?(o|a)mm?a)?$", "i");
 	Classes.fetch("https://api.yomomma.info/").then((reply: string) => command.data.cache.push(reply));