+++ lib/typings/lib/Classes.d.ts
@@ -4,13 +4,16 @@ import * as adm_panel from "adm-panel2";
 import { EventEmitter } from "events";
 import { RequestOptions } from "https";
 import { URL } from "url";
+import * as chillout from "chillout";
 export declare const chalk: any;
 export declare var stripAnsi: {
     (c: any): any;
 };
 /**
- * VAL-1: TO BE USED BOTH WITH USER AND BOTS ACCOUNTS
- * VAL-2: RELOAD ONLY MODIFIED MODULES/COMMANDS!
+ * VAL-1: TO BE USED BOTH WITH USER AND BOTS ACCOUNTS  - can make botnet
+ * VAL-2: RELOAD ONLY MODIFIED MODULES/COMMANDS!  - external handling, _loadCMD supports singlefile
+ * VAL-3: CLEAR UNDERLYING LOGSTRINGS PERIODICALLY
+ * VAL-4: CACHEBANK AND NAMEDCACHEBANK IMPLEMENTATION
  */
 export declare module Classes {
     namespace Options {
@@ -67,8 +70,10 @@ export declare module Classes {
         constructor(opts?: Options.ValeOpts);
         on(event: "log", listener: (...args: any[]) => void): this;
         on(event: "rawlog", listener: (...args: any[]) => void): this;
+        once(event: "log", listener: (...args: any[]) => void): this;
+        once(event: "rawlog", listener: (...args: any[]) => void): this;
         start(): this;
-        command(message: Discord.Message): Promise<string | void>;
+        command(message: Discord.Message): Promise<any>;
         _debug(...msg: any): string;
         _loadCMD(from?: string): Promise<this>;
     }
@@ -80,8 +85,8 @@ export declare module Classes {
         category?: string;
         data?: any;
         constructor(opts: Options.CommandOpts);
-        body(message?: Discord.Message, vale?: Vale): Promise<void>;
-        _remove(vale?: Vale): Promise<void>;
+        body(message?: Discord.Message, vale?: Vale): Promise<any>;
+        _remove(vale?: Vale): Promise<any>;
     }
     class CacheBank implements Options.CacheBankOpts {
         size: number;
@@ -92,7 +97,7 @@ export declare module Classes {
         private static cntr;
         constructor(name?: string, size?: number, autopurge?: boolean, reusables?: boolean);
         get(item: number): any;
-        purge(items?: number): CacheEntry[];
+        purge(items?: number): Promise<CacheEntry[]>;
         push(item: any): number;
         _arrange(): CacheEntry[];
     }
@@ -100,4 +105,5 @@ export declare module Classes {
     function failsafe(this: Discord.Message, ...params: any[]): Promise<Discord.Message | Discord.Message[]>;
 }
 export default Classes;
+export { chillout };
 //# sourceMappingURL=Classes.d.ts.map