+++ CHANGELOG.md
@@ -1,5 +1,15 @@
   
-# Changelog 2019-05-29T21:14:21.700Z  
+# Changelog  
+  
+## 2019-05-31T16:12:23.098Z  
+  
+- `!urban word\<String>`  
+  
+## 2019-05-30T11:05:19.069Z  
+  
+- Updated underlying panel.  
+  
+## 2019-05-29T21:14:21.700Z  
   
 - Caching implemented, speeding up external API requests.  
 - Command reloading leaks patched.